Établissement des requêtes HTTP vers un API Météo externe

+++ app/src/main/java/dti/g55/eventich_client/domaine/entite/ConditionMeterologique.kt
@@ -1,4 +1,4 @@
 package dti.g55.eventich_client.domaine.entite
 
-class ConditionMeterologique( val meteoDescription: String, val températureMoyenne: Int, val pourcentageHumidité: Int, val listeHeures: ArrayList<HeureMeteo> ) {
+class ConditionMeterologique( val meteoDescription: String, val températureMoyenne: Double, val pourcentageHumidité: Int, val listeHeures: ArrayList<HeureMeteo> ) {
 }
